% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fseq_env.R
\name{fseq_env}
\alias{fseq_env}
\alias{fseq_get_env}
\alias{fseq_print_env}
\alias{fseq_set_env}
\alias{fseq_check_env}
\alias{fseq_copy_fun}
\alias{fun_var_env}
\title{Alter the environments of fseq functions}
\usage{
fseq_get_env(fn_fseq, i)

fseq_print_env(fn_fseq, i)

fseq_set_env(fn_fseq, i, new_env)

fseq_check_env(new_env)

fseq_copy_fun(fn_fseq)

fun_var_env(fun, variable, value)
}
\arguments{
\item{fn_fseq}{a function of class \code{fseq} (with an appropriate environment).}

\item{i}{a numeric indicating which function in the \code{fseq} to refer to. If
0, then the parent \code{fseq} environment will be used.}

\item{new_env}{environment to be assigned to the relevant environment in \code{fn_fseq}.}

\item{fun}{function to edit the environment of.}

\item{variable}{name of variable (as string) to assign value to within all
environments in \code{fun}.}

\item{value}{value to be assigned to \code{variable} in all environments in \code{fun}}
}
\description{
These functions allows access to the environments contained
within \code{fseq} objects. \code{fseq} objects are functions that have been
created using the pipeline with a \code{.} as the left-hand side.

\code{fseq} functions contain a major environment for the overall
function (i.e the environment of \code{fn_fseq}) and multiple
minor environments. The minor environments are the
function-environments of each of functions that make up the
functional sequence.

For example, if we define \code{a <- . \%>\% add(2) \%>\% multiply_by(3)},
then \code{a()} will be a function that has an environment, this is the
major environment. Within \code{a()}, there is a list of functions that
make up the function sequence, \code{add(.,2)} and \code{multiply_by(.,3)}.
Each of these will also have their own environment, the minor
environments. Therefore, \code{a()} will have 3 environments associated
with it.
}
