% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if-pipe.R
\name{if-pipe}
\alias{if-pipe}
\alias{if_pipe}
\title{Function to branch within a pipeline}
\usage{
if_pipe(.x, .predicate, .fun, .elsefun = NULL)
}
\arguments{
\item{.x}{Input data that is being fed through the pipeline}

\item{.predicate}{Logical value or vector of logicals indicating whether
`.fun` or `.elsefun` should be ran}

\item{.fun}{Pipeline or list of pipeline functions to be ran if .predicate evaluates to `TRUE`}

\item{.elsefun}{Pipeline or list of pipeline functions to be ran if .predicate evaluates to `FALSE`
If `NULL`, no function is applied when `.predicate == FALSE`}
}
\description{
Provides the ability to branch your pipeline without breaking it
}
\examples{
iris \%>\%
  if_pipe(T,
          mutate(Sepal.Area = Sepal.Length*Sepal.Width) \%>\%
            head,
          summarise(Sepal.Area = sum(Sepal.Length*Sepal.Width)))

iris \%>\%
  if_pipe(F,
          mutate(Sepal.Area = Sepal.Length*Sepal.Width) \%>\%
            head,
          summarise(Sepal.Area = sum(Sepal.Length*Sepal.Width)))

}
