% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpipe.R
\name{mpipe}
\alias{mpipe}
\title{The mpipe package}
\description{
The \code{mpipe} package is designed as an extension to
the normal pipe format of coding. It introduces new, key features
and allows more users to perform actions that require you
to break out of a pipeline much less.
}
