% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe-func.R
\name{pipe-func}
\alias{pipe-func}
\alias{pipe_func}
\title{Run a pipe as a function}
\usage{
pipe_func(lhs, rhs)
}
\arguments{
\item{lhs}{The left hand side of the pipe}

\item{rhs}{The right had side of the pipe}
}
\value{
evaluates the pipe `lhs %>% rhs`
}
\description{
Function to run a part of a pipeline in function notation,
 a wrapper around `%>%`
}
\details{
Converts the input `lhs` and `rhs` as call strings,
concatenates them with the general pipe
}
