% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fseq_methods.R
\name{fseq_methods}
\alias{fseq_methods}
\alias{+.fseq}
\alias{length.fseq}
\alias{is.fseq}
\alias{is_fseq}
\title{Generic methods for fseq manipulations}
\usage{
\method{+}{fseq}(lhs, rhs)

\method{length}{fseq}(x)

is.fseq(object)

is_fseq(object)
}
\arguments{
\item{lhs}{\code{fseq} to be performed first}

\item{rhs}{\code{fseq} to be performed second}

\item{x}{an \code{fseq} object.}

\item{object}{object to be tested.}
}
\description{
When pipelines are created without an initial term, the
\code{magrittr} package (where the pipes originate)
creates a specialised type of function called a \code{fseq}.
However, \code{magrittr} doesn't do much with these
\code{fseq} objects.
}
\examples{
a <- . \%>\%
  magrittr::add(2) \%>\%
  magrittr::divide_by(3)
b <- . \%>\%
  rep(3) \%>\%
  sum()

(a + b)(2)
(b + a)(2)
length(a)
is.fseq(a)
is.fseq(mean)
is_fseq(a)
is_fseq(mean)
}
