% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe_cat.R
\name{pipe_cat}
\alias{pipe_cat}
\title{Pipe-able version of the \code{cat()}}
\usage{
pipe_cat(
  data,
  ...,
  file = "",
  sep = " ",
  fill = FALSE,
  labels = NULL,
  append = FALSE
)
}
\arguments{
\item{data}{the data being passed through the pipeline}

\item{...}{arguments to be passed to the \code{cat()} function. Arguments will be
evaluated in the context of \code{data} before being passed to \code{cat()}}

\item{file, sep, fill, labels, append}{See the \code{\link[base:cat]{cat()}} documentation for more
information regarding these arguments.}
}
\description{
The \code{pipe_cat()} function allows messages to be output to
the console or to an external file without the need to break out
of a pipeline to do so.
}
\examples{

sample(100, 1) \%>\%
  runif() \%>\%
  pipe_cat("Current length: ", length(.), "\n") \%>\%
  pipe_cat("Current average: ", mean(.), "\n") \%>\%
  pipe_cat("Current standard error: ", sd(.) / length(.), "\n") \%>\%
  pipe_cat("Returning mean:\n") \%>\%
  mean()

tibble::tibble(
  x = runif(10),
  y = runif(10)
) \%>\%
  pipe_cat("Average x: ", mean(x), "\n") \%>\%
  pipe_cat("Current number of rows: ", nrow, "\n") \%>\%
  dplyr::mutate(z = x + y)

palmerpenguins::penguins \%>\%
  dplyr::mutate(species = as.character(species)) \%>\%
  dplyr::filter(!is.na(bill_length_mm)) \%>\%
  pipe_cat("Total average Culmen Length: ", mean(bill_length_mm), "\n") \%>\%
  dplyr::group_by(species) \%>\%
  pipe_cat(species, " average Culmen Length: ", mean(bill_length_mm), "\n")
}
